# apps/web/src/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerEvaluationResponse": {
        "description": "Owner-scoped evaluation record without source-content duplication.",
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message_id": {
            "title": "Message Id",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "rating": {
            "enum": [
              "supported",
              "incomplete",
              "incorrect"
            ],
            "title": "Rating",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "message_id",
          "rating",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "AnswerEvaluationResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Stable error information nested inside the API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "examples": [
              "VALIDATION_ERROR",
              "HTTP_404",
              "INTERNAL_ERROR"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional field-level validation details.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable error message.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Common error envelope returned by every API error handler.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "description": "Correlation ID for this request.",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvaluationListResponse": {
        "description": "Recent answer evaluations and their aggregate counts.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnswerEvaluationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          },
          "summary": {
            "$ref": "#/components/schemas/EvaluationSummaryResponse"
          }
        },
        "required": [
          "items",
          "page",
          "summary"
        ],
        "title": "EvaluationListResponse",
        "type": "object"
      },
      "EvaluationSummaryResponse": {
        "description": "Aggregate grounded-answer evaluation counts.",
        "properties": {
          "incomplete": {
            "minimum": 0.0,
            "title": "Incomplete",
            "type": "integer"
          },
          "incorrect": {
            "minimum": 0.0,
            "title": "Incorrect",
            "type": "integer"
          },
          "supported": {
            "minimum": 0.0,
            "title": "Supported",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "supported",
          "incomplete",
          "incorrect"
        ],
        "title": "EvaluationSummaryResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageMetadata": {
        "description": "Metadata returned with paginated collection responses.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "minimum": 0.0,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "PageMetadata",
        "type": "object"
      },
...
      "SavedSearchCreateRequest": {
        "description": "Input for a reusable search definition.",
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "filters": {
            "additionalProperties": true,
            "title": "Filters",
            "type": "object"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "query": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "name",
          "query"
        ],
        "title": "SavedSearchCreateRequest",
        "type": "object"
      },
      "SavedSearchResponse": {
        "description": "Owner-scoped reusable hybrid-search definition.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "filters": {
            "additionalProperties": true,
            "title": "Filters",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "query",
          "filters",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "title": "SavedSearchResponse",
        "type": "object"
      },
      "SavedSearchUpdateRequest": {
        "description": "Partial update for a reusable search definition.",
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "query": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          }
        },
        "title": "SavedSearchUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationErrorDetail": {
        "description": "A machine-readable validation problem location.",
        "properties": {
          "location": {
            "description": "Path to the invalid request value.",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "description": "Human-readable validation message.",
            "title": "Message",
            "type": "string"
          },
          "type": {
            "description": "Pydantic validation error type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ValidationErrorDetail",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP API for the self-hosted ThoughtHarbor second brain.",
    "title": "ThoughtHarbor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evaluations": {
      "get": {
        "operationId": "list_evaluations_api_v1_evaluations_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evaluations",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/saved-searches": {
      "get": {
        "operationId": "list_saved_searches_api_v1_saved_searches_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedSearchResponse"
                  },
                  "title": "Response List Saved Searches Api V1 Saved Searches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          }
        },
        "summary": "List Saved Searches",
        "tags": [
          "review"
        ]
      },
      "post": {
        "operationId": "create_saved_search_api_v1_saved_searches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Saved Search",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/saved-searches/{search_id}": {
      "delete": {
        "operationId": "delete_saved_search_api_v1_saved_searches__search_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "search_id",
            "required": true,
            "schema": {
              "title": "Search Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Saved Search",
        "tags": [
          "review"
        ]
      },
      "patch": {
        "operationId": "update_saved_search_api_v1_saved_searches__search_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "search_id",
            "required": true,
            "schema": {
              "title": "Search Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Saved Search",
        "tags": [
          "review"
        ]
      }
    },
...
  },
...
}